Make png code been suitable to use old and new png apis
The new API ( post libpng12 ) is default on SYSTEM_LIBPNG usage
The old API is used if internal 3rdparty code is used.

Keep the original oldapi was a port choice due to avoid update the
internal 3rdparty sources and still keeping the possibility of compile
internal 3rdparty sources need to be kept for now to verify
compatibility. For now only internal libpng is enabled.